OPAL-238:   - Upgraded to POI version 3.7 to pick up fix for https://issues.apache.org/bugzilla/show_bug.cgi?id=48936.   - ExcelDatasource: Removed code that was added to workaround this bug.   - ExcelDatasourceTest: Updated OPAL-238 test (renamed as "test_OPAL_238_strings_can_be_written"). This test passes if the test strings can be written without error (which they can be now with POI 3.7; they could not be with POI 3.6).
NOTE: This does NOT fix the problem described in an earlier comment, which occurs when creating an ExcelDatasource then disposing of it (by either removing it or by shutting down Opal). A crash occurs, which may be related to a bug in xml-beans: https://issues.apache.org/bugzilla/show_bug.cgi?id=49940.